+++ src/main/java/com/luckyframe/project/testmanagmt/projectProtocolTemplate/controller/ProjectTemplateParamsController.java
@@ -13,6 +13,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
 
+import com.luckyframe.common.utils.security.PermissionUtils;
 import com.luckyframe.framework.aspectj.lang.annotation.Log;
 import com.luckyframe.framework.aspectj.lang.enums.BusinessType;
 import com.luckyframe.framework.web.controller.BaseController;
@@ -111,7 +112,12 @@ public class ProjectTemplateParamsController extends BaseController
 	@RequestMapping(value = "/editSave",method=RequestMethod.POST,consumes="application/json")
 	@ResponseBody
 	public AjaxResult editSave(@RequestBody List<ProjectTemplateParams> projectTemplateParams)
-	{
+	{		
+		if(!PermissionUtils.isProjectPermsPassByProjectId(projectProtocolTemplateService
+				.selectProjectProtocolTemplateById(projectTemplateParams.get(0).getTemplateId()).getProjectId())){
+			return error("没有此项目保存模板参数权限");
+		}
+		
 		int result=0;
 		projectTemplateParamsService.deleteProjectTemplateParamsByIds(projectTemplateParams);
 		for(ProjectTemplateParams projectTemplateParam:projectTemplateParams){
